requirements.txt
pandas==2.3.3
numpy==2.3.5
matplotlib==3.10.7
seaborn==0.13.2
mysql-connector-python==9.1.0
python-dotenv==1.0.0
